nodex v0.1.0 Nodex

Nodex is a OTP Application which will take care of your connected nodes.

Nodex can launch Nodex.NodesWatchers which will be the responsibles of taking care of the connected nodes. By default there is a Nodex.NodesWatcher already launched.

start a Nodex.NodesWatcher

iex> Nodex.start_nodes_watcher(:my_nodes_watcher)

and stop it later:

iex> Nodex.stop_nodes_watcher(:my_nodes_watcher)

In order to connect to a node we only have to do the following:

iex> Nodex.connect(nodo, :my_nodes_watcher) #specific nodes_watcher
iex> Nodex.connect(nodo) #default nodes_watcher

A node is specified using Nodex.Node and the simplest example would be:

iex> Nodex.connect(%Nodex.Node{app_name: :app, host_address: "my.host"})

Functions

connect(nodo, nodes_watcher \\ :nodes_watcher_default)

It connects to a node using an active Nodex.NodesWatcher. If no Nodex.NodesWatcher is given, it will used the default one. The node has to be specified using a Nodex.Node.

start_nodes_watcher(name)

It starts a new Nodex.NodesWatcher. The input has to be a term() because it will be used as name of the process.

stop_nodes_watcher(name)

It stops the Nodex.NodesWatcher whose name is given.