nodex v0.1.0 Nodex.NodesWatcher

Nodex.NodesWatcher is the a GenServer which will be in charge of supervising its connected nodes and try to reconnect if the node is down.

By Default there is a default Nodex.NodesWatcher but it is possible to start different Nodex.NodesWatcher in order to have better structure.

iex> Nodex.NodesWatcher.start_link(%Nodex.NodesWatcher{name: :my_nodes_watcher})

To connect to a node you have to:

iex> Nodex.connect(%Nodex.Node{app_name: :app, host_address: "my.host"}, node_watcher)
# or Using the default nodes watcher
iex> Nodex.connect(%Nodex.Node{app_name: :app, host_address: "my.host"})

When you need to do execute anything in the remote code just after the connection:

iex> remote_procedure = fn(node_name) -> my_function(node_name) end
iex> Nodex.connect(%Nodex.Node{app_name: :app, host_address: "my.host", fun: remote_procedure})

A reconnection time can be specified in Nodex.Node:

iex> %Nodex.Node{app_name: :app, host_address: "my.host", reconnection_time: 10_000}

Functions

connect(node_arguments, nodes_watcher \\ :nodes_watcher_default)

It will use the given Nodex.NodesWatcher in order handle the connection to the given Nodex.Node

start_link(nodes_watcher)